Fixes circular dependency problems
Instead of depending directly on the `computeTextAlternative` function, we create a type representing the function and pass it down as an argument to rules that require it.

This allows us to move every rule's code to the appropriate file, which make it much easier to find when debugging unit tests.

This change also makes the context mandatory for all rules that use it. It introduces a `createRuleRunner` test utility to automatically provide the `computeTextAlternative` function and a default context when needed. This allows us to preserve the current shape of the tests.

PiperOrigin-RevId: 338674875