Ensures that accname does not remove too many spaces
Most of the steps assumed that the result should be a trimmed flat string, but doing so in intermediate steps can cause some important spaces to be dropped.

This change ensures that this assumption is removed and that the name is only trimmed at the very end, to preserve all of the spacing information.

PiperOrigin-RevId: 349256970